refactor(react-gui): give the density map services a folder
Density maps were spread over four `.service.ts` files and two more in
`helpers/`, none of which said they belonged together: probing a file's
header before reading it, streaming it in, choosing MTZ columns, and the
461-line panel that lists a scene's map renderers and edits them.

`map/` collects all of it and splits the panel by what its calls do:
`renderers.ts` finds the map renderers in a scene, `state.ts` reads one as
the panel row shows it, `props.ts` writes them. `redrawMapCenter` sits with
the writes because it is the same edit seen from the other side -- the
displayed region follows the view, so moving the camera has to move the
map's centre for the contour to stay where the user is looking.

The three small services keep their code and lose their `.service.ts`
suffix (`probeHeader`, `streamLoad`, `mtzColumns`), since the folder now has
one registry entry. `helpers/setupDensityMapRenderers` and
`helpers/emMapDefaults` move in too -- only the map path used them.

No behaviour change; the existing map tests pass with only their import
paths moved.
