+++ tritium/react-gui/src/renderer/__test__/emMapDefaults.test.ts
@@ -11,7 +11,7 @@ import {
     fitViewsToMap,
     isEmDensityMap,
     EM_INITIAL_TOP_FRACTION,
-} from '@renderer/worker/server/services/helpers/emMapDefaults'
+} from '@renderer/worker/server/services/map/emDefaults'
 import type { FormatOptions } from '@renderer/dialogs/fopen-opt-dlgs/types'
 
 function ccp4Format(mapType: 'auto' | 'xtal' | 'em'): FormatOptions {
