+++ tritium/react-gui/src/renderer/__test__/getMtzColumnInfoService.test.ts
@@ -8,7 +8,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import type { WorkerContext } from '@renderer/worker/server/types/WorkerContext'
 
-import { services } from '@renderer/worker/server/services/getMtzColumnInfo.service'
+import { services } from '@renderer/worker/server/services/map/map.service'
 const { getMtzColumnInfo } = services
 
 function makeCtx(opts: {
